requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langchain-core>=0.2.0
langchain-groq>=0.1.5
rank-bm25>=0.2.2
requests>=2.31.0
beautifulsoup4>=4.12.0
jinja2>=3.1.3
python-dotenv>=1.0.1
python-multipart>=0.0.9
pytest>=9.0.0
google-api-python-client>=2.193.0
google-auth>=2.49.1
